Removing dll function, add check for process open
Removes the `InjectDllIntoProcess` function, which was unused. Added a check for whether `FindProcessId` actually found the target process, so that the program exits without any changes if it's run while okami.exe is not open. Finally, I re-added the `CloseHandle(process);` line at the end, which I had accidentally removed at some point.