Adding fix for Ryo music being silenced by brush
Previously, with the mod enabled, using the brush in Ryoshima Coast and North Ryoshima Coast would cause the music to go silent. I've added a patch which prevents this and lets the Ryo music continue playing. I'm not 100% sure how it works, but it just NOPs out a function which seems to be associated with the gradual volume change of the background music when pulling out the brush. The one downside is that if you are in Ryo, pull out the brush, and *then* enable the mod, the music will be permanently silenced until you reload the area (I think), even if you disable the mod.